% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialize_connection}
\alias{initialize_connection}
\title{Initialisation de la connexion à la base de données.}
\usage{
initialize_connection()
}
\value{
dbConnection Connexion à la base de données oracle
}
\description{
Si le code s'exécute en dehors du portail, initie une connexion sqlite pour
effectuer les tests.
}
